python-services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.51.0
anthropic==0.34.0
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
sympy==1.13.3
numpy==1.26.4
requests==2.32.3
httpx==0.27.2
supabase==2.8.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
anyio==4.6.0
slowapi==0.1.9
PyMuPDF==1.24.3